requirements.txt
python~=3.9
django~=4.2.7
djangorestframework~=3.14.0